Moved Drag'n'Drop components to own folder/renamed them
